requirements.txt
Flask
flask-cors
flask-swagger-ui
pydantic
python-dotenv
bleach
pandas
xgboost